refactor code to process_bgen, change output logic
